requirements.txt
playwright==1.41.0
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pytz==2023.3
requests==2.31.0
